+++ lib/event-notifications/index.ts
@@ -1,1 +1,3 @@
 export * from "./eventNotificationPipeline";
+
+export * from "./emitClient";